Revert "got rid of responsiveness classes on home page and depend on conditional rendering" restored old naming of components This reverts commit 08965b52e50a98655e3697da85e75c72d7f8a97a.
